requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pyodbc==5.2.0
python-dotenv==1.0.1
